+++ CHANGES.rst
@@ -39,10 +39,10 @@ Changes
 - The row indices of training and testing samples are now also included in the
   dictionaries produced by :meth:`DataOp.skb.iter_cv_splits`. :pr:`2012` by
   :user:`Jérôme Dockès <jeromedockes>`.
-- The :class:`Cleaner` now exposes a ``parse_strings`` boolean parameter to
+- The :class:`Cleaner` now exposes a ``parse_numbers`` boolean parameter to
   control whether numeric-looking strings (e.g., ``["1", "2", "3"]``) are parsed
-  to ``np.float32``, and a ``cast_to_float`` parameter to downcast floating-point
-  columns to ``np.float32`` (without converting integer columns).
+  to ``np.float32``, and a ``cast_to_float`` parameter to downcast numeric
+  columns to ``np.float32``.
   :pr:`1910` by :user:`<Varshith-yadaV>`.
 - :func:`fetch_toxicity_dataset` now returns a shuffled version of the dataset by default.
   :pr:`1892` by user:`Riccardo Cappuzzo <rcap107>`.